requirements.txt
deflate>=0.3.0
Pillow>=8.4.0
scikit-image>=0.19.2